# docs/openapi.yaml
openapi: 3.1.0
info:
  title: gs-webhook-2-events
  version: 1.0.0
  description: |
    Webhook receiver for Grand Shooting events, publishing to gs-stream-api.

    This service receives webhooks from Grand Shooting when pictures or references are created, updated, or deleted.
    Events are then published to the gs-stream-api for further processing by other services.

    ## Authentication

    ### Webhook Endpoint
    The webhook endpoint requires a per-client secret key passed as a query parameter (`key`).
    Each client has its own unique `webhookSecretKey` that is automatically generated when the client is created.
    The key can be regenerated at any time via the admin API.

    ### Admin API
    The admin API requires an API key passed in the `x-api-key` header.

    ## Event Types
    The following event types are supported:
    - `pictures/create` - Picture created
    - `pictures/update` - Picture updated
    - `pictures/delete` - Picture deleted
    - `references/create` - Reference created
    - `references/update` - Reference updated
    - `references/delete` - Reference deleted

  contact:
    name: Grand Shooting
    url: https://grand-shooting.com
    email: help@grand-shooting.com
  license:
    name: Private
    identifier: LicenseRef-Proprietary
servers:
  - url: https://gs-webhook-2-events.grand-shooting.com
    description: Production
  - url: https://gs-webhook-2-events-staging.grand-shooting.com
    description: Staging
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      name: x-api-key
      in: header
      description: Admin API key for accessing administration endpoints
  schemas:
...
    WebhookPayload:
      type: object
      required:
        - account_id
        - topic
        - type
        - ids
      properties:
        account_id:
          type: integer
          description: Grand Shooting account ID
        topic:
          type: string
          description: Event topic (e.g., pictures/create)
        type:
          type: string
          enum: [picture, reference]
          description: Resource type
        ids:
          type: array
          items:
            oneOf:
              - type: integer
              - type: string
          description: List of affected resource IDs
        data:
          type: object
          additionalProperties: true
          description: Resource data keyed by ID
        picture_id:
          type: array
          items:
            type: integer
          description: Deprecated - use ids instead
          deprecated: true
    Error:
      type: object
      properties:
        error:
          type: string
paths:
...
  /webhook/{accountId}:
    post:
      tags:
        - Webhook
      summary: Receive webhook events
      description: |
        Receive webhook events from Grand Shooting.

        This endpoint is called by Grand Shooting when pictures or references are created, updated, or deleted.
        The service validates the payload, checks if the event type is enabled for the client, and publishes
        the event to gs-stream-api.
      operationId: receiveWebhook
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
          description: Grand Shooting account ID
        - name: key
          in: query
          required: true
          schema:
            type: string
          description: Per-client secret key for authentication (obtain from admin API)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPayload'
            example:
              account_id: 12345
              topic: pictures/create
              type: picture
              ids: [1, 2, 3]
              data:
                '1':
                  name: photo1.jpg
                '2':
                  name: photo2.jpg
                '3':
                  name: photo3.jpg
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  processed:
                    type: integer
                    description: Number of events processed
                  failed:
                    type: integer
                    description: Number of events that failed
        '400':
          description: Invalid payload or parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid secret key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...